moved session/login related code to separate "session" folder
