+++ app/app.module.ts
@@ -3,8 +3,8 @@ import {BrowserModule} from '@angular/platform-browser';
 
 import {AppComponent} from './app.component';
 import {routing} from "./app.routes";
-import {SessionService} from "./user/session.service";
-import {LoggedInGuard} from "./user/logged-in.guard";
+import {SessionService} from "./session/session.service";
+import {LoggedInGuard} from "./session/logged-in.guard";
 import {EntityMapperService} from "./model/entity-mapper.service";
 import {databaseServiceProvider, DatabaseManagerService} from "./database/database-manager.service";
 import {PouchDatabaseManagerService} from "./database/pouch-database-manager.service";
@@ -12,7 +12,7 @@ import {NavigationItemsService} from "./navigation/navigation-items.service";
 import {ConfigService} from "./config/config.service";
 import {FormsModule} from "@angular/forms";
 import {SyncStatusComponent} from "./sync-status/sync-status.component";
-import {LoginComponent} from "./user/login.component";
+import {LoginComponent} from "./session/login.component";
 import {FooterComponent} from "./footer.component";
 import {NavigationComponent} from "./navigation/navigation.component";
 import {AlertsModule} from "./alerts/alerts.module";
