+++ app/app.routes.ts
@@ -1,7 +1,7 @@
 import {Routes, RouterModule}  from '@angular/router';
 import {DashboardComponent} from "./dashboard/dashboard.component";
 import {UserAccountComponent} from "./user/user-account.component";
-import {LoggedInGuard} from "./user/logged-in.guard";
+import {LoggedInGuard} from "./session/logged-in.guard";
 import {ModuleWithProviders} from "@angular/core";
