+++ app/session/session.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-import { User } from "./user";
+import { User } from "../user/user";
 import { DatabaseManagerService } from "../database/database-manager.service";
 import { AlertService } from "../alerts/alert.service";
 import { EntityMapperService } from "../model/entity-mapper.service";
